Both tank2 and tank3 use TempCalcMethod.SimpleBeta as default
The Tank2Cfg (for the 2-pico TankModule2) had SimpleBetaForPico as
its default. At 60F this was reporting ~40F, as it was assuming
the pico had an impedance of 30 kOhms in the voltage divider formula.

This change allows us to uniformly clean up all the temp sense reporting
to use the same SimpleBeta formula